Adjust FT update and delete

+++ components/director/internal/domain/formationtemplate/repository_test.go
@@ -45,13 +45,13 @@ func TestRepository_Get(t *testing.T) {
 }
 
 func TestRepository_GetByNameAndTenant(t *testing.T) {
-	suite := testdb.RepoGetTestSuite{
-		Name:       "Get Formation Template By name and tenant",
+	suiteWithTenant := testdb.RepoGetTestSuite{
+		Name:       "Get Formation Template By name and tenant when tenant is present",
 		MethodName: "GetByNameAndTenant",
 		SQLQueryDetails: []testdb.SQLQueryDetails{
 			{
-				Query:    regexp.QuoteMeta(`SELECT id, name, application_types, runtime_types, runtime_type_display_name, runtime_artifact_kind, tenant_id FROM public.formation_templates WHERE name = $1 AND tenant_id = $2`),
-				Args:     []driver.Value{formationTemplateName, testTenantID},
+				Query:    regexp.QuoteMeta(`SELECT id, name, application_types, runtime_types, runtime_type_display_name, runtime_artifact_kind, tenant_id FROM public.formation_templates WHERE tenant_id = $1 AND name = $2`),
+				Args:     []driver.Value{testTenantID, formationTemplateName},
 				IsSelect: true,
 				ValidRowsProvider: func() []*sqlmock.Rows {
 					return []*sqlmock.Rows{sqlmock.NewRows(fixColumns()).AddRow(formationTemplateEntity.ID, formationTemplateEntity.Name, formationTemplateEntity.ApplicationTypes, formationTemplateEntity.RuntimeTypes, formationTemplateEntity.RuntimeTypeDisplayName, formationTemplateEntity.RuntimeArtifactKind, formationTemplateEntity.TenantID)}
@@ -71,7 +71,34 @@ func TestRepository_GetByNameAndTenant(t *testing.T) {
 		DisableConverterErrorTest: false,
 	}
 
-	suite.Run(t)
+	suiteWithoutTenant := testdb.RepoGetTestSuite{
+		Name:       "Get Formation Template By name and tenant when tenant is not present",
+		MethodName: "GetByNameAndTenant",
+		SQLQueryDetails: []testdb.SQLQueryDetails{
+			{
+				Query:    regexp.QuoteMeta(`SELECT id, name, application_types, runtime_types, runtime_type_display_name, runtime_artifact_kind, tenant_id FROM public.formation_templates WHERE name = $1 AND tenant_id IS NULL`),
+				Args:     []driver.Value{formationTemplateName},
+				IsSelect: true,
+				ValidRowsProvider: func() []*sqlmock.Rows {
+					return []*sqlmock.Rows{sqlmock.NewRows(fixColumns()).AddRow(formationTemplateEntity.ID, formationTemplateEntity.Name, formationTemplateEntity.ApplicationTypes, formationTemplateEntity.RuntimeTypes, formationTemplateEntity.RuntimeTypeDisplayName, formationTemplateEntity.RuntimeArtifactKind, formationTemplateEntityNullTenant.TenantID)}
+				},
+				InvalidRowsProvider: func() []*sqlmock.Rows {
+					return []*sqlmock.Rows{sqlmock.NewRows(fixColumns())}
+				},
+			},
+		},
+		ConverterMockProvider: func() testdb.Mock {
+			return &automock.EntityConverter{}
+		},
+		RepoConstructorFunc:       formationtemplate.NewRepository,
+		ExpectedModelEntity:       &formationTemplateModelNullTenant,
+		ExpectedDBEntity:          &formationTemplateEntityNullTenant,
+		MethodArgs:                []interface{}{formationTemplateName, ""},
+		DisableConverterErrorTest: false,
+	}
+
+	suiteWithoutTenant.Run(t)
+	suiteWithTenant.Run(t)
 }
 
 func TestRepository_Create(t *testing.T) {
@@ -129,14 +156,14 @@ func TestRepository_Exists(t *testing.T) {
 }
 
 func TestRepository_Delete(t *testing.T) {
-	suite := testdb.RepoDeleteTestSuite{
-		Name: "Delete Formation Template By ID",
+	suiteWithoutTenant := testdb.RepoDeleteTestSuite{
+		Name: "Delete Formation Template By ID when there is no tenant",
 		SQLQueryDetails: []testdb.SQLQueryDetails{
 			{
-				Query:         regexp.QuoteMeta(`DELETE FROM public.formation_templates WHERE id = $1`),
+				Query:         regexp.QuoteMeta(`DELETE FROM public.formation_templates WHERE id = $1 AND tenant_id IS NULL`),
 				Args:          []driver.Value{testID},
 				ValidResult:   sqlmock.NewResult(-1, 1),
-				InvalidResult: sqlmock.NewResult(-1, 0),
+				InvalidResult: sqlmock.NewResult(-1, 2),
 			},
 		},
 		RepoConstructorFunc: formationtemplate.NewRepository,
@@ -144,10 +171,28 @@ func TestRepository_Delete(t *testing.T) {
 			return &automock.EntityConverter{}
 		},
 		IsGlobal:   true,
-		MethodArgs: []interface{}{testID},
+		MethodArgs: []interface{}{testID, ""},
 	}
 
-	suite.Run(t)
+	suiteWithTenant := testdb.RepoDeleteTestSuite{
+		Name: "Delete Formation Template By ID when there is tenant",
+		SQLQueryDetails: []testdb.SQLQueryDetails{
+			{
+				Query:         regexp.QuoteMeta(`DELETE FROM public.formation_templates WHERE tenant_id = $1 AND id = $2`),
+				Args:          []driver.Value{testTenantID, testID},
+				ValidResult:   sqlmock.NewResult(-1, 1),
+				InvalidResult: sqlmock.NewResult(-1, 2),
+			},
+		},
+		RepoConstructorFunc: formationtemplate.NewRepository,
+		ConverterMockProvider: func() testdb.Mock {
+			return &automock.EntityConverter{}
+		},
+		MethodArgs: []interface{}{testID, testTenantID},
+	}
+
+	suiteWithTenant.Run(t)
+	suiteWithoutTenant.Run(t)
 }
 
 func TestRepository_List(t *testing.T) {
@@ -247,13 +292,34 @@ func TestRepository_List(t *testing.T) {
 }
 
 func TestRepository_Update(t *testing.T) {
-	updateStmt := regexp.QuoteMeta(`UPDATE public.formation_templates SET name = ?, application_types = ?, runtime_types = ?, runtime_type_display_name = ?, runtime_artifact_kind = ? WHERE id = ?`)
-	suite := testdb.RepoUpdateTestSuite{
-		Name: "Update Formation Template By ID",
+	updateStmtWithoutTenant := regexp.QuoteMeta(`UPDATE public.formation_templates SET name = ?, application_types = ?, runtime_types = ?, runtime_type_display_name = ?, runtime_artifact_kind = ? WHERE id = ?`)
+	suiteWithoutTenant := testdb.RepoUpdateTestSuite{
+		Name: "Update Formation Template By ID without tenant",
 		SQLQueryDetails: []testdb.SQLQueryDetails{
 			{
-				Query:         updateStmt,
+				Query:         updateStmtWithoutTenant,
 				Args:          []driver.Value{formationTemplateEntity.Name, formationTemplateEntity.ApplicationTypes, formationTemplateEntity.RuntimeTypes, formationTemplateEntity.RuntimeTypeDisplayName, formationTemplateEntity.RuntimeArtifactKind, formationTemplateEntity.ID},
+				ValidResult:   sqlmock.NewResult(-1, 1),
+				InvalidResult: sqlmock.NewResult(-1, 0),
+			},
+		},
+		RepoConstructorFunc: formationtemplate.NewRepository,
+		ConverterMockProvider: func() testdb.Mock {
+			return &automock.EntityConverter{}
+		},
+		ModelEntity:    &formationTemplateModelNullTenant,
+		DBEntity:       &formationTemplateEntityNullTenant,
+		NilModelEntity: nilModelEntity,
+		IsGlobal:       true,
+	}
+
+	updateStmtWithTenant := regexp.QuoteMeta(`UPDATE public.formation_templates SET name = ?, application_types = ?, runtime_types = ?, runtime_type_display_name = ?, runtime_artifact_kind = ? WHERE id = ? AND tenant_id = ?`)
+	suiteWithTenant := testdb.RepoUpdateTestSuite{
+		Name: "Update Formation Template By ID with tenant",
+		SQLQueryDetails: []testdb.SQLQueryDetails{
+			{
+				Query:         updateStmtWithTenant,
+				Args:          []driver.Value{formationTemplateEntity.Name, formationTemplateEntity.ApplicationTypes, formationTemplateEntity.RuntimeTypes, formationTemplateEntity.RuntimeTypeDisplayName, formationTemplateEntity.RuntimeArtifactKind, formationTemplateEntity.ID, formationTemplateEntity.TenantID},
 				ValidResult:   sqlmock.NewResult(-1, 1),
 				InvalidResult: sqlmock.NewResult(-1, 0),
 			},
@@ -265,8 +331,8 @@ func TestRepository_Update(t *testing.T) {
 		ModelEntity:    &formationTemplateModel,
 		DBEntity:       &formationTemplateEntity,
 		NilModelEntity: nilModelEntity,
-		IsGlobal:       true,
 	}
 
-	suite.Run(t)
+	suiteWithTenant.Run(t)
+	suiteWithoutTenant.Run(t)
 }
